# docs/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: Cybertech E-Commerce API
  description: |
    Backend REST API for the Cybertech e-commerce platform.

    Covers the full e-commerce lifecycle: product catalog, search,
    cart, wishlist, orders, payments (Stripe), reviews and moderation,
    bank cards (PCI-aware), discount campaigns, user management
    (Keycloak-backed) and behavioural event ingestion.

    Authentication is performed via Keycloak-issued JWT bearer tokens.
    Public endpoints (catalog browsing, registration, Stripe webhook)
    are explicitly whitelisted; every other endpoint requires a
    valid Bearer token and may further restrict to USER or ADMIN roles.
  contact:
    name: Cybertech Team
    email: dev@cybertech.local
  license:
    name: Proprietary
    url: https://cybertech.local/license
  version: "1.0"
servers:
- url: http://localhost:8081
  description: Local dev
- url: https://api.cybertech.local
  description: Production
security:
- keycloak: []
...
paths:
...
  /api/v1/services/wishlist/add/{productUuid}:
    post:
      tags:
      - Wishlist
      summary: Add product to my wishlist
      description: Adds a specific product to the connected user's wishlist.
      operationId: addProductToWishlist
      parameters:
      - name: productUuid
        in: path
        description: UUID of the product to add
        required: true
        schema:
          type: string
          format: uuid
      - name: X-API-VERSION
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
      responses:
        "201":
          description: Product added successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistResponseDto"
        "401":
          description: Unauthorized - User not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: Forbidden - Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "409":
          description: Conflict - Product already in wishlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
      - keycloak: []
...
  /api/v1/services/wishlist/my-wishlist:
    get:
      tags:
      - Wishlist
      summary: Get my wishlist
      description: Retrieves all products in the connected user's wishlist.
      operationId: getMyWishlist
      parameters:
      - name: X-API-VERSION
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
      responses:
        "200":
          description: Wishlist retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WishlistResponseDto"
        "401":
          description: Unauthorized - User not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: Forbidden - Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
      - keycloak: []
...
  /api/v1/services/wishlist/remove/{productUuid}:
    delete:
      tags:
      - Wishlist
      summary: Remove product from my wishlist
      description: Removes a specific product from the connected user's wishlist.
      operationId: removeProductFromWishlist
      parameters:
      - name: productUuid
        in: path
        description: UUID of the product to remove
        required: true
        schema:
          type: string
          format: uuid
      - name: X-API-VERSION
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
      responses:
        "204":
          description: Product removed successfully
        "401":
          description: Unauthorized - User not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: Forbidden - Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "404":
          description: Product not in wishlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
      - keycloak: []
...
components:
  schemas:
...
    ErrorResponseDto:
      type: object
      properties:
        message:
          type: string
        httpStatusCode:
          type: integer
          format: int32
        errorCodeType:
          type: string
          enum:
          - TECHNICAL
          - FUNCTIONAL
    ProductResponseDto:
      type: object
      properties:
        name:
          type: string
        uuid:
          type: string
        price:
          type: number
        brand:
          type: string
        category:
          type: string
        photoUrls:
          type: array
          items:
            type: string
        description:
          type: string
        attributes:
          type: object
          additionalProperties: {}
    WishlistResponseDto:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        product:
          $ref: "#/components/schemas/ProductResponseDto"
        addedAt:
          type: string
          format: date-time
...
  securitySchemes:
    keycloak:
      type: http
      description: "Keycloak-issued JWT access token. Obtain one from the realm token\
        \ endpoint and pass it as the Authorization header: 'Bearer <token>'."
      scheme: bearer
      bearerFormat: JWT